billdesksysback/requirements.txt
fastapi==0.115.6
loguru==0.7.2
pandas==2.2.3
passlib==1.7.4
pydantic==2.10.3
python_jose==3.3.0
redis==5.2.0
SQLAlchemy==2.0.32
user_agents==2.2.0
uvicorn==0.32.1
